moved bstats code from Main to BStatsManager and added debug for every chart
